* 'master' of https://github.com/JMRI/JMRI: (174 commits)
  ESU LokSound v3.x main-files
  ESU LokSound v3.x optimization and translation
  More LayoutSlipTest @Test cases
  Added more LayoutTurnoutTest cases
  Refactor pp.drawControls; javadoc fix for Path.java
  PositionablePont: Added test cases to increase code coverage
  Update help files.  add some debug statements
  SectionTransitBundle Dutch translation Fix Transit Special Action spinner values after change of What Some Dutch spelling fixes
  new test
  new test; remove unnecessary graphics check
  rename test file to match class under test; new test
  rename test to match class under test
  new tests
  new test
  new tests
  correctly initialize programmer variable
  new tests
  remove unused import
  more Junit4 conversions of PackageTests
  convert to JUnit4 and Jemmy
  ...

+++ java/src/jmri/managers/DefaultSignalMastLogicManager.java
@@ -338,15 +338,15 @@ public class DefaultSignalMastLogicManager implements jmri.SignalMastLogicManage
         throw new UnsupportedOperationException("Not supported yet.");
     }
 
-    long signalLogicDelay = 500L;
+    int signalLogicDelay = 500;
 
     @Override
-    public long getSignalLogicDelay() {
+    public int getSignalLogicDelay() {
         return signalLogicDelay;
     }
 
     @Override
-    public void setSignalLogicDelay(long l) {
+    public void setSignalLogicDelay(int l) {
         signalLogicDelay = l;
     }
